INSPECTIT-2221: Moving/Refactoring InfluxDbDao to another package
