+++ inspectit.server/src/test/java/rocks/inspectit/server/processor/impl/InfluxProcessorTest.java
@@ -20,7 +20,7 @@ import org.mockito.Mock;
 import org.testng.annotations.Test;
 
 import rocks.inspectit.server.influx.builder.DefaultDataPointBuilder;
-import rocks.inspectit.server.influx.dao.impl.InfluxDBDao;
+import rocks.inspectit.server.influx.dao.InfluxDBDao;
 import rocks.inspectit.shared.all.communication.DefaultData;
 import rocks.inspectit.shared.all.communication.data.HttpTimerData;
 import rocks.inspectit.shared.all.communication.data.InvocationSequenceData;
